refactor(gateway): centralize startup tracing helpers
